renames resource package
