+++ service/controller/v3/resource/networkconfig/create.go
@@ -1,4 +1,4 @@
-package networkconfigv2
+package networkconfig
 
 import (
 	"context"
