+++ service/controller/v3/resource/networkconfig/resource.go
@@ -1,4 +1,4 @@
-package networkconfigv2
+package networkconfig
 
 import (
 	"github.com/giantswarm/microerror"
