+++ service/controller/v3/resource/networkconfig/spec.go
@@ -1,4 +1,4 @@
-package networkconfigv2
+package networkconfig
 
 type Backend struct {
 	Type string
